it-24.3.21374.21517

+++ BaseApp/Source/Base Application/Manufacturing/Journal/ProductionJournalMgt.Codeunit.al
@@ -45,35 +45,37 @@ codeunit 5510 "Production Journal Mgt"
         LeaveForm: Boolean;
         IsHandled: Boolean;
     begin
-        OnBeforeHandling(ProdOrder);
-
-        MfgSetup.Get();
-
-        SetTemplateAndBatchName();
-
-        InitSetupValues();
-
-        ProgressBar.Open(GeneratingJnlLinesMsg);
-        DeleteJnlLines(ToTemplateName, ToBatchName, ProdOrder."No.", ActualLineNo);
-        CreateJnlLines(ProdOrder, ActualLineNo);
-        ProgressBar.Close();
-
         IsHandled := false;
-        OnBeforeRunProductionJnl(ToTemplateName, ToBatchName, ProdOrder, ActualLineNo, PostingDate, IsHandled);
+        OnBeforeHandling(ProdOrder, ActualLineNo, IsHandled);
         if not IsHandled then begin
-            repeat
-                // Commit before running Production Journal page
-                Commit();
+            MfgSetup.Get();
 
-                LeaveForm := true;
-                Clear(ProductionJnl);
-                ProductionJnl.Setup(ToTemplateName, ToBatchName, ProdOrder, ActualLineNo, PostingDate);
-                ProductionJnl.RunModal();
-                if DataHasChanged(ToTemplateName, ToBatchName, ProdOrder."No.", ActualLineNo) then
-                    LeaveForm := Confirm(Text001, true);
-            until LeaveForm;
+            SetTemplateAndBatchName();
 
+            InitSetupValues();
+
+            ProgressBar.Open(GeneratingJnlLinesMsg);
             DeleteJnlLines(ToTemplateName, ToBatchName, ProdOrder."No.", ActualLineNo);
+            CreateJnlLines(ProdOrder, ActualLineNo);
+            ProgressBar.Close();
+
+            IsHandled := false;
+            OnBeforeRunProductionJnl(ToTemplateName, ToBatchName, ProdOrder, ActualLineNo, PostingDate, IsHandled);
+            if not IsHandled then begin
+                repeat
+                    // Commit before running Production Journal page
+                    Commit();
+
+                    LeaveForm := true;
+                    Clear(ProductionJnl);
+                    ProductionJnl.Setup(ToTemplateName, ToBatchName, ProdOrder, ActualLineNo, PostingDate);
+                    ProductionJnl.RunModal();
+                    if DataHasChanged(ToTemplateName, ToBatchName, ProdOrder."No.", ActualLineNo) then
+                        LeaveForm := Confirm(Text001, true);
+                until LeaveForm;
+
+                DeleteJnlLines(ToTemplateName, ToBatchName, ProdOrder."No.", ActualLineNo);
+            end;
         end;
     end;
 
@@ -558,7 +560,13 @@ codeunit 5510 "Production Journal Mgt"
         ItemJnlLine2: Record "Item Journal Line";
         ReservEntry: Record "Reservation Entry";
         HasChanged: Boolean;
+        IsHandled: Boolean;
     begin
+        IsHandled := false;
+        OnBeforeDataHasChanged(TemplateName, BatchName, ProdOrderNo, ProdOrderLineNo, HasChanged, IsHandled);
+        if IsHandled then
+            exit(HasChanged);
+
         ItemJnlLine2.Reset();
         ItemJnlLine2.SetRange("Journal Template Name", TemplateName);
         ItemJnlLine2.SetRange("Journal Batch Name", BatchName);
@@ -665,7 +673,7 @@ codeunit 5510 "Production Journal Mgt"
     end;
 
     [IntegrationEvent(false, false)]
-    local procedure OnBeforeHandling(ProdOrder: Record "Production Order")
+    local procedure OnBeforeHandling(ProdOrder: Record "Production Order"; ActualLineNo: Integer; var IsHandled: Boolean)
     begin
     end;
 
@@ -781,6 +789,11 @@ codeunit 5510 "Production Journal Mgt"
 
     [IntegrationEvent(false, false)]
     local procedure OnDeleteJnlLinesOnAfterSetFilters(var ItemJournalLine: Record "Item Journal Line")
+    begin
+    end;
+
+    [IntegrationEvent(false, false)]
+    local procedure OnBeforeDataHasChanged(TemplateName: Code[10]; BatchName: Code[10]; ProdOrderNo: Code[20]; ProdOrderLineNo: Integer; var HasChanged: Boolean; var IsHandled: Boolean)
     begin
     end;
 }
